+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 
+- Implemented `__str__` for FaultTree
+
 
 ## [v0.7.2] PyPI tweaks (2024-06-14)